+++ lib/pin_ui.dart
@@ -1,5 +1,3 @@
-library pin_ui;
-
 export 'src/indicator/animation_controller.dart';
 export 'src/indicator/exceptions/controller_not_initialized_exception.dart';
 export 'src/indicator/models/implementations.dart'